turns Functor into a factory object, reducing class size
