+++ lib/streamer/functors.rb
@@ -1,0 +1,19 @@
+require 'streamer/functors/functor'
+require 'streamer/functors/divide'
+require 'streamer/functors/eq'
+require 'streamer/functors/group'
+require 'streamer/functors/gt'
+require 'streamer/functors/gte'
+require 'streamer/functors/least'
+require 'streamer/functors/lookup'
+require 'streamer/functors/lt'
+require 'streamer/functors/lte'
+require 'streamer/functors/member'
+require 'streamer/functors/multiply'
+require 'streamer/functors/subtract'
+require 'streamer/functors/sum'
+module Streamer
+  # Functors supply the ability to perform functions on a payload
+  module Functors
+  end
+end
