+++ lib/streamer/functors/eq.rb
@@ -1,0 +1,10 @@
+module Streamer
+  module Functors
+    # Eq compares target to function or value
+    class Eq < Functor
+      def call
+        compare :==
+      end
+    end
+  end
+end
