+++ lib/streamer/functors/gt.rb
@@ -1,0 +1,10 @@
+module Streamer
+  module Functors
+    # Gt compares target to function or value
+    class Gt < Functor
+      def call
+        compare :>
+      end
+    end
+  end
+end
