+++ lib/streamer/functors/gte.rb
@@ -1,0 +1,10 @@
+module Streamer
+  module Functors
+    # Gte compares a target to a value or fuction
+    class Gte < Functor
+      def call
+        compare(:>=)
+      end
+    end
+  end
+end
