+++ lib/streamer/functors/lte.rb
@@ -1,0 +1,10 @@
+module Streamer
+  module Functors
+    # Lte compares target to function or value
+    class Lte < Functor
+      def call
+        compare :<=
+      end
+    end
+  end
+end
